+++ api/routers/filer/dependencies/parameters.py
@@ -3,21 +3,22 @@ from typing import Annotated
 from sqlalchemy.ext.asyncio import AsyncSession
 from aiohttp import ClientSession
 
+from api.config.settings import get_settings
 from api.common.enums import ResponseFormat
 from api.common.formatters import clean, print_enum_values
 from api.dependencies.database import DatabaseSessionManager
+from api.dependencies.http_client import HttpClientSessionManager
 from api.dependencies.parameters.optional import get_response_format
 from api.dependencies.parameters.services import InternalRequestParameters as BaseInternalRequestParameters
 
 from ..common.constants import ROUTE_DATABASE
-from .api_session import FILERApiSessionManager
 
 # override session to use the ROUTE_SESSION_MANAGER
 ROUTE_SESSION_MANAGER = DatabaseSessionManager(ROUTE_DATABASE)
-API_SESSION_MANAGER = FILERApiSessionManager()
+API_CLIENT_SESSION_MANAGER = HttpClientSessionManager(get_settings().FILER_REQUEST_URI)
 class InternalRequestParameters(BaseInternalRequestParameters, arbitrary_types_allowed=True):
     session: Annotated[AsyncSession, Depends(ROUTE_SESSION_MANAGER)]
-    apiWrapper: Annotated[ClientSession, Depends(API_SESSION_MANAGER)]
+    apiClientSession: Annotated[ClientSession, Depends(API_CLIENT_SESSION_MANAGER)]
 
 async def path_track_id(track: str = Path(description="FILER track identifier")) -> str:
     return clean(track)
